Remove generate cmd from project
