+++ datumaro/util/telemetry_utils.py
@@ -58,6 +58,8 @@ def _get_action_name(command):
         return "diff_result"
     elif command is commands.explain.explain_command:
         return "explain_result"
+    elif command is commands.generate.generate_command:
+        return "generate_result"
     elif command is commands.info.info_command:
         return "info_result"
     elif command is commands.log.log_command:
