+++ tests/cli/test_generator.py
@@ -1,5 +1,4 @@
 from unittest import TestCase
-
 import os
 import os.path as osp
